+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* target Android 15
 * Improve readability of info messages in dark mode
 
 ## v2.11.0